suzerain: silence unsigend signed warning
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@18752 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/channel_mean.cpp
@@ -100,7 +100,7 @@ static bool process(const char * filename)
     load(h.get(), scenario);
     load(h.get(), grid);
     load(h.get(), bspw);
-    assert(bspw->ndof() == grid.Ny);
+    assert(static_cast<unsigned>(bspw->ndof()) == grid.Ny);
 
     // Load zero-zero mode coefficients for all state variables
     matrix_t<>::real s_coeffs(grid.Ny, field_names.static_size);
